Create enumearable list
